Moved scanSymbolReferencesInJar to ClassDumper
