+++ dependencies/src/main/java/com/google/cloud/tools/opensource/classpath/ClassDumper.java
@@ -19,6 +19,7 @@ package com.google.cloud.tools.opensource.classpath;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Preconditions;
 import com.google.common.base.Predicates;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
@@ -31,6 +32,7 @@ import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.net.URLClassLoader;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.security.CodeSource;
@@ -146,7 +148,31 @@ class ClassDumper {
     return methodReferences.build();
   }
 
-  static SymbolReferenceSet scanSymbolReferencesInClass(JavaClass javaClass) {
+  /**
+   * Scans class files in the jar file and returns a {@link SymbolReferenceSet} populated with
+   * symbolic references.
+   *
+   * @param jarFilePath absolute path to a jar file
+   * @return symbol references and classes defined in the jar file
+   */
+  static SymbolReferenceSet scanSymbolReferencesInJar(Path jarFilePath) {
+    Preconditions.checkArgument(
+        jarFilePath.isAbsolute(), "The input jar file path is not an absolute path");
+    Preconditions.checkArgument(
+        Files.isReadable(jarFilePath), "The input jar file path is not readable");
+
+    SymbolReferenceSet.Builder symbolTableBuilder = SymbolReferenceSet.builder();
+    try {
+      for (JavaClass javaClass : topLevelJavaClassesInJar(jarFilePath)) {
+        symbolTableBuilder.merge(scanSymbolReferencesInClass(javaClass));
+      }
+      return symbolTableBuilder.build();
+    } catch (ClassNotFoundException | IOException ex) {
+      throw new RuntimeException("Failed to scan jar file", ex);
+    }
+  }
+
+  private static SymbolReferenceSet scanSymbolReferencesInClass(JavaClass javaClass) {
     SymbolReferenceSet.Builder symbolTableBuilder = SymbolReferenceSet.builder();
     ImmutableSet.Builder<MethodSymbolReference> methodReferences =
         symbolTableBuilder.methodReferencesBuilder();
